+++ server/AccTelemetryTracker.Logic/IDiscordNotifier.cs
@@ -6,6 +6,7 @@ public interface IDiscordNotifier
     /// </summary>
     /// <param name="motecFile">The motec file object that was just upload</param>
     /// <param name="avatarUrl">The avatar of the discord user</param>
+    /// <param name="anyFasterLaps">If there are any existing motec files that contain a faster lap than this new one</param>
     /// <returns></returns>
-    Task Notify(Datastore.Models.MotecFile motecFile, string? avatarUrl);
+    Task Notify(Datastore.Models.MotecFile motecFile, string? avatarUrl, bool anyFasterLaps);
 }